update avater docs
