+++ src/documentation/examples/Avatar/Size/Small.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import Avatar from '@bufferapp/ui/Avatar';
+
+/** Small */
+export default function ExampleAvatar() {
+  return (
+    <Avatar
+      src="https://pbs.twimg.com/profile_images/988613046510628866/Io1ZQUpy_400x400.jpg"
+      alt="@joelgascoigne"
+      size="small"
+    />
+  );
+}
