+++ src/documentation/examples/Avatar/Type/AvatarSocial.jsx
